+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/pisdf/pimm2sdf/StaticPiMM2SDFLauncher.java
@@ -81,7 +81,7 @@ public class StaticPiMM2SDFLauncher {
     SDFGraph result;
 
     // Get all the available values for all the parameters
-    final Map<Parameter, Integer> parametersValues = getParametersValues();
+    final Map<Parameter, Long> parametersValues = getParametersValues();
 
     // Visitor creating the SDFGraph
     final PiGraphExecution execution = new PiGraphExecution(parametersValues);
@@ -97,8 +97,8 @@ public class StaticPiMM2SDFLauncher {
    *
    * @return the parameters values
    */
-  private Map<Parameter, Integer> getParametersValues() {
-    final Map<Parameter, Integer> result = new LinkedHashMap<>();
+  private Map<Parameter, Long> getParametersValues() {
+    final Map<Parameter, Long> result = new LinkedHashMap<>();
 
     final Set<Entry<Parameter, String>> parameterValues = this.scenario.getParameterValues().entrySet();
     for (final Entry<Parameter, String> paramValue : parameterValues) {
@@ -110,11 +110,11 @@ public class StaticPiMM2SDFLauncher {
             + " transformation (id: org.ietr.preesm.experiment.pimm2sdf.PiMM2SDFTask)");
       } else {
         try {
-          final int value = Integer.parseInt(paramValue.getValue());
+          final long value = Long.parseLong(paramValue.getValue());
           result.put(paramValue.getKey(), value);
         } catch (final NumberFormatException e) {
           // The expression associated to the parameter is an
-          // expression (and not an constant int value).
+          // expression (and not an constant long value).
           // Leave it as it is, it will be solved later.
         }
       }
